Real Life Raspberry PI demo!
A new version of the demo in my other repository RasPI-NTC-ADC, now I'm
using my library! Stuck the Thermistor in my mouth during the test, you
can see the temperature rise :-) It's a bit flaky in the measurements,
I think it's the ADC that for some reason looses accuracy. Have Fun!